backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
cdp-sdk==0.0.6
python-dotenv==0.19.0